refact: 계산 로직을 features로 분리
